Reorganizes the code
